+++ resale/src/main/java/com/goti/resale/service/domain/ResaleHoldExpiryDomainService.java
@@ -1,6 +1,7 @@
 package com.goti.resale.service.domain;
 
 import java.time.LocalDateTime;
+import java.util.UUID;
 
 import com.goti.resale.domain.entity.resale.ResaleHoldEntity;
 import com.goti.resale.domain.entity.resale.ResaleListingEntity;
@@ -11,4 +12,6 @@ public interface ResaleHoldExpiryDomainService {
 		ResaleHoldEntity resaleHold,
 		LocalDateTime now
 	);
+
+	void expire(UUID holdId, LocalDateTime now);
 }
