+++ resale/src/main/java/com/goti/resale/service/domain/ResaleHoldExpiryDomainServiceImpl.java
@@ -1,15 +1,27 @@
 package com.goti.resale.service.domain;
 
 import java.time.LocalDateTime;
+import java.util.UUID;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.goti.constants.messages.ErrorCode;
+import com.goti.exception.CustomException;
 import com.goti.resale.domain.entity.resale.ResaleHoldEntity;
 import com.goti.resale.domain.entity.resale.ResaleListingEntity;
 import com.goti.global.validation.Preconditions;
+import com.goti.resale.repository.hold.ResaleHoldRepository;
+import com.goti.resale.repository.listing.ResaleListingRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class ResaleHoldExpiryDomainServiceImpl implements ResaleHoldExpiryDomainService {
+
+	private final ResaleHoldRepository resaleHoldRepository;
+	private final ResaleListingRepository resaleListingRepository;
 
 	@Override
 	public void expire(
@@ -27,5 +39,21 @@ public class ResaleHoldExpiryDomainServiceImpl implements ResaleHoldExpiryDomain
 
 		resaleListing.releaseHold();
 		resaleHold.release();
+	}
+
+	@Override
+	@Transactional
+	public void expire(UUID holdId, LocalDateTime now) {
+		ResaleHoldEntity resaleHold = resaleHoldRepository.findById(holdId)
+			.orElseThrow(
+				() -> new CustomException(ErrorCode.RESALE_HOLD_NOT_FOUND)
+			);
+
+		ResaleListingEntity resaleListing = resaleHold.getResaleListing();
+
+		expire(resaleListing, resaleHold, now);
+
+		resaleHoldRepository.save(resaleHold);
+		resaleListingRepository.save(resaleListing);
 	}
 }
